Customize defaults in models
In model itself. Still needs to allow using a different primary key name

+++ db/migrations/20170127143149_create_users.cr
@@ -1,7 +1,7 @@
 class CreateUsers::V20170127143149 < Avram::Migrator::Migration::V1
   def migrate
     create :users do
-      primary_key id : Int32
+      primary_key id : Int64
       add_timestamps
       add name : String
       add nickname : String?
