+++ db/migrations/20170127143150_create_admins.cr
@@ -1,7 +1,7 @@
 class CreateAdmins::V20170127143150 < Avram::Migrator::Migration::V1
   def migrate
     create :admins do
-      primary_key id : Int32
+      primary_key id : Int64
       add_timestamps
       add name : String
     end
